requirements.txt
asgiref==3.8.1
Django==5.1.3
django-allauth==65.3.0
django-cors-headers==4.6.0
django-rest-auth==0.9.5
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
PyJWT==2.10.1
six==1.16.0
sqlparse==0.5.2
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
